+++ functions/Config.fnc.php
@@ -8,17 +8,21 @@
 
 /**
  * Get Configuration value
+ * Insert or update (for current school) value if passed as argument.
  *
  * @example  Config( 'SYEAR' )
+ *
+ * @since 4.4 Add $value param to INSERT or UPDATE.
  *
  * @global array  $_ROSARIO     Sets $_ROSARIO['Config']
  * @global string $DefaultSyear
  *
- * @param  string $item         Config title.
+ * @param  string $item  Config title.
+ * @param  string $value Value to INSERT or UPDATE. Defaults to null.
  *
  * @return string Config value
  */
-function Config( $item )
+function Config( $item, $value = null )
 {
 	global $_ROSARIO,
 		$DefaultSyear;
@@ -31,21 +35,38 @@ function Config( $item )
 	// Get General & School Config.
 	if ( ! isset( $_ROSARIO['Config'][ (string) $item ] ) )
 	{
-		// General (for every school) Config is stored with SCHOOL_ID=0.
-		$school_where = "SCHOOL_ID='0'";
+		$school_where = UserSchool() > 0 ?
+			// If user logged in.
+			"SCHOOL_ID='" . UserSchool() . "' OR SCHOOL_ID='0' ORDER BY SCHOOL_ID DESC" :
+			// General (for every school) Config is stored with SCHOOL_ID=0.
+			"SCHOOL_ID='0'";
 
-		// If user logged in.
-		if ( UserSchool() > 0 )
-		{
-			$school_where = "SCHOOL_ID='" . UserSchool() . "' OR " . $school_where .
-				" ORDER BY SCHOOL_ID DESC";
-		}
-
-		$_ROSARIO['Config'] = DBGet( DBQuery( "SELECT TITLE, CONFIG_VALUE
+		$_ROSARIO['Config'] = DBGet( DBQuery( "SELECT TITLE,CONFIG_VALUE,SCHOOL_ID
 			FROM CONFIG
 			WHERE " . $school_where ), array(), array( 'TITLE' ) );
 
 		$_ROSARIO['Config']['SYEAR'][1]['CONFIG_VALUE'] = $DefaultSyear;
+	}
+
+	if ( ! is_null( $value ) )
+	{
+		if ( ! isset( $_ROSARIO['Config'][ (string) $item ][1]['CONFIG_VALUE'] ) )
+		{
+			// Insert value (does not exist).
+			DBQuery( "INSERT INTO CONFIG (CONFIG_VALUE,TITLE,SCHOOL_ID)
+				VALUES('" . $value . "','" . $item . "','" .
+				( UserSchool() > 0 ? UserSchool() : '0' ) . "'" );
+		}
+		elseif ( $value !== $_ROSARIO['Config'][ (string) $item ][1]['CONFIG_VALUE'] )
+		{
+			// Update value (different from current value).
+			DBQuery( "UPDATE CONFIG
+				SET CONFIG_VALUE='" . $value . "'
+				WHERE TITLE='" . $item . "'
+				AND SCHOOL_ID='" . $_ROSARIO['Config'][ (string) $item ][1]['SCHOOL_ID'] . "'" );
+		}
+
+		$_ROSARIO['Config'][ (string) $item ][1]['CONFIG_VALUE'] = $value;
 	}
 
 	return $_ROSARIO['Config'][ (string) $item ][1]['CONFIG_VALUE'];
@@ -56,21 +77,24 @@ function Config( $item )
  * Get Program Configuration
  * Get 1 value if item specified,
  * else get Program values
+ * Insert or update value if passed as argument.
  *
  * Values set in School Configuration or directly in Module (ex.: Eligibility Entry times)
  *
  * @example if ( ProgramConfig( 'students', 'STUDENTS_SEMESTER_COMMENTS' ) )
  *
  * @since 2.9
+ * @since 4.4 Add $value param to INSERT or UPDATE.
  *
  * @global array        $_ROSARIO Sets $_ROSARIO['ProgramConfig']
  *
  * @param  string $program eligibility|grades|students|moodle|food_service|attendance... Program name.
  * @param  string $item    Program Config title (optional). Defaults to 'all'.
+ * @param  string $value   Value to INSERT or UPDATE. Defaults to null.
  *
  * @return string|array Program Configuration value, or Program values in array
  */
-function ProgramConfig( $program, $item = 'all'  )
+function ProgramConfig( $program, $item = 'all', $value = null )
 {
 	global $_ROSARIO;
 
@@ -86,6 +110,29 @@ function ProgramConfig( $program, $item = 'all'  )
 			FROM PROGRAM_CONFIG
 			WHERE SYEAR='" . UserSyear() . "'
 			AND SCHOOL_ID='" . UserSchool() . "'" ), array(), array( 'PROGRAM', 'TITLE' ) );
+	}
+
+	if ( ! is_null( $value )
+		&& $item !== 'all' )
+	{
+		if ( ! isset( $_ROSARIO['ProgramConfig'][ (string) $program ][ (string) $item ][1]['VALUE'] ) )
+		{
+			// Insert value (does not exist).
+			DBQuery( "INSERT INTO PROGRAM_CONFIG (VALUE,PROGRAM,TITLE,SCHOOL_ID,SYEAR)
+				VALUES('" . $value . "','" . $program . "','" . $item . "','" .
+				UserSchool() . "','" . UserSyear() . "'" );
+		}
+		elseif ( $value !== $_ROSARIO['ProgramConfig'][ (string) $program ][ (string) $item ][1]['VALUE'] )
+		{
+			// Update value (different from current value).
+			DBQuery( "UPDATE PROGRAM_CONFIG
+				SET VALUE='" . $value . "'
+				WHERE TITLE='" . $item . "'
+				AND SCHOOL_ID='" . UserSchool() . "'
+				AND SYEAR='" . UserSyear() . "'" );
+		}
+
+		$_ROSARIO['ProgramConfig'][ (string) $program ][ (string) $item ][1]['VALUE'] = $value;
 	}
 
 	if ( $item === 'all' )
@@ -106,6 +153,7 @@ function ProgramConfig( $program, $item = 'all'  )
  * Program User Config
  * To get all config options at once
  * If you want only one option, prefer `Preferences()`
+ * Insert or update value if passed as argument.
  *
  * @example $gradebook_config = ProgramUserConfig( 'Gradebook' );
  *
@@ -113,13 +161,15 @@ function ProgramConfig( $program, $item = 'all'  )
  * @see PROGRAM_USER_CONFIG table
  *
  * @since 2.9
+ * @since 4.4 Add $value param to INSERT or UPDATE.
  *
  * @param string  $program  Gradebook|WidgetsSearch|StaffWidgetsSearch|
  * @param integer $staff_id Staff ID (optional). Defaults to User( 'STAFF_ID' ).
+ * @param string $value     Value to INSERT or UPDATE. Defaults to null.
  *
  * @return array Program User Config, associative array( '[title]' => '[value]' ).
  */
-function ProgramUserConfig( $program, $staff_id = 0 )
+function ProgramUserConfig( $program, $staff_id = 0, $value = null )
 {
 	static $program_config;
 
@@ -130,20 +180,41 @@ function ProgramUserConfig( $program, $staff_id = 0 )
 
 	$staff_id = $staff_id ? $staff_id : User( 'STAFF_ID' );
 
-	$config_RET = DBGet( DBQuery( "SELECT TITLE,VALUE
-		FROM PROGRAM_USER_CONFIG
-		WHERE USER_ID='" . $staff_id . "'
-		AND PROGRAM='" . $program . "'" ), array(), array( 'TITLE' ) );
-
-	if ( $config_RET )
+	if ( ! isset( $program_config[ $program ][ $staff_id ] ) )
 	{
+		$config_RET = DBGet( DBQuery( "SELECT TITLE,VALUE
+			FROM PROGRAM_USER_CONFIG
+			WHERE USER_ID='" . $staff_id . "'
+			AND PROGRAM='" . $program . "'" ), array(), array( 'TITLE' ) );
+
+		$program_config[ $program ][ $staff_id ] = null;
+
 		foreach ( (array) $config_RET as $title => $value )
 		{
 			$program_config[ $program ][ $staff_id ][ $title ] = $value[1]['VALUE'];
 		}
 	}
-	else
-		$program_config[ $program ][ $staff_id ] = null;
+
+	if ( ! is_null( $value ) )
+	{
+		if ( ! isset( $program_config[ $program ][ $staff_id ][ $title ] ) )
+		{
+			// Insert value (does not exist).
+			DBQuery( "INSERT INTO PROGRAM_USER_CONFIG (VALUE,PROGRAM,TITLE,USER_ID)
+				VALUES('" . $value . "','" . $program . "','" . $title . "','" .
+				$staff_id . "'" );
+		}
+		elseif ( $value !== $program_config[ $program ][ $staff_id ][ $title ] )
+		{
+			// Update value (different from current value).
+			DBQuery( "UPDATE PROGRAM_USER_CONFIG
+				SET VALUE='" . $value . "'
+				WHERE TITLE='" . $title . "'
+				AND USER_ID='" . UserSchool() . "'" );
+		}
+
+		$program_config[ $program ][ $staff_id ][ $title ] = $value;
+	}
 
 	return $program_config[ $program ][ $staff_id ];
 }